2021-D3B Use contextual filtering to find the right bits
Basically using a tree to count prefixes - i.e. the count of a tree's
node is the number of entries that have the same prefix as you would use
to navigate to that node from the root. So the root node is the total
number (since all entries have the empty prefix). The root's first and
second children has the number of entries that start with 1 and 0,
respectively, and so on.

This works, but turns out to be a bit awkward to then traverse, since
you're effectively rebuilding the prefix, which requires knowledge of
the children at that level (e.g. at the root node, to know if the first
element in the prefix is a 1 or a 0 you need to check how big the
children are, and possibly if they even exist or are just leaves).