2021-D4B Find the last winning game rather than the first
For the first solution I used the Either monad to capture the first
winning game and propagate that through the whole process. Obviously
that doesn't work to find the last winning game. Luckily however there
is another monad that could work: Writer. Basically as each game is won,
I append it to the log of completed games (along with the winning
number, which is required for the score calculation), and also remove it
from the active games in play. So at the end of the process the last
game in the log is the last game that was won.