2021-D6B Counting a lot more laternfish
So as alluded to in the previous commit, the new goal of calculating the
fish population 256 days in the future is too much for the naive
approach. Something a bit more efficient is needed.

Taking a look at a single fish, the number of fish it will be an
ancestor to after a given number of days can be expressed as a
recurrence formula. And what's a great way to efficiently compute
recurrence formulae? Dynamic programming! And as it turns out, Haskell's
laziness makes this rather easy, assuming you know the trick to it. You
simply define a list or array or whatever recursively, following the
recurrence relation. The list is lazy, so any particular value won't be
computed until called for, but it also won't be computed more than once.
And so you get an efficient solution that doesn't look much different
than the naive implementation. Yay!