2021-D7B Modified fuel calculations
Turns out the brute force approach was fine for this part as well. In
fact, given how the "distance" (i.e. fuel) calculation changed, finding
a more efficient approach to narrow down the possibilities to try for
the first solution might not have worked for the second solution at all
so it would have been a waste of time.

Instead the trick in the second part (or at least /a/ trick) is knowing
the identity that the sum of the first n integers equals (n*(n+1)/2),
which makes finding the new solution no more or less efficient than the
first solution (since they're both constant time to calculate the fuel
cost of a single crab to a single position).