2021-D8A Count easily identifiable digits
Based on signals to a digital readout where the normal wiring would be

 aaaa
b    c
b    c
 dddd
e    f
e    f
 gggg

But the wiring has been scrambled (the scrambling is unique in each line
of input). The input is a series of lines with the 10 unique
combinations corresponding to each (scrambled) digit (in unknown order),
a pipe, and the desired output (using the same cipher as the patterns in
the first half of the line).

The first task is to simply count the number of digits in the output
that can be uniquely identified by number of wires alone.