2021-D8B Fully decoding the mixed wires
There may be a more elegant way to solve this, but the first thing that
sprung to my mind was to see if the number of digits a given wire
(letter) appeared in would be helpful, and it was. Of the 7 wires, 3 can
be identified by the number of digits they appear in, and the other 4
are paired up in a way that lets me identify them using the method from
part one. e.g., both a and c appear in 8 digits, but only c appears in
the combination for "1" (which uniquely uses two wires). Figuring out
how to code this was a bit difficult, and I don't think it turned out
especially elegant, but it was still fairly easy to do.