2021-D10B Completionism
Good thing I didn't define a data type to hold the results of the
parsing, because it turned out to be totally unnecessary. I didn't even
need to build up the suffix to complete the valid chunks because of
course that would be the reversed equivalent of the remaining stack from
parsing. And since the non-error lines result in a Right containing that
stack, I can easily reuse that code, yay.

Also, getting the middle element of a list is just annoying.