2021-D11A Flashing octopuses
This one was kind of awkward. I used Either to keep track of whether an
octopus had already flashed in a step (not sure if it's even possible
for one to flash twice, but wanted to be sure). Unfortunately wasn't
able to find a way to use its monadic properties to simplify anything,
but it still worked fine.

I was able to use the list monad to generate a point's adjacent points
in a fairly readable way, so that's nice.