2021-D11B Find the point where the flashes all sync up
A fairly simple recursion, though it felt kind of awkward. In the first
part I had considered trying to use the State monad to keep track of the
number of flashes, but that didn't really seem worth it. Now I think it
might have helped reduce the awkwardness of this recursion. Oh well.