2021-D12A Searching caves
So this is pretty obviously a graph. However, the readily findable Graph
type did not seem like the right fit, as that deals with vertices of
Ints. It is possible to associate those vertices with some other "key"
value, but in the end for this question it seemed like too many hoops to
jump through. I knew I would have to implement my own traversal anyway
to allow for visiting "big" caves multiple times, so I just used
a node-neighbour map.