2021-D12B Revisiting your favourite small cave
The main challenge here was that I discovered a bug in my previous
solution: the traverseCaves filtering by isValidStep was being done on
the path passed into the function, which does not include the current
node. I guess it worked in the previous one because no node has a
self-edge, so checking if a node's neighbours are in the path doesn't
depend on the node being in the path. But it does matter now, since it's
possible for two repeating small caves in a row to occur, and the new
flag wouldn't be set properly when considering the second one.

Speaking of the flag, it seemed like the easiest way to keep track of
whether a path already has a repeated small cave in it. Otherwise you'd
have to search the entire path for duped small caves, since the one you
are considering might not be the one that was already duped.