2021-D13B Printing the dots
This is the first problem with a non-numerical answer, which is cool.

I wanted to have the printing logic be as simple as possible, so I
wanted to go row by row. To that end, I was thinking about sorting the
dot points to be in that order, then filling in the gaps, so I replaced
the normal (Int, Int) tuple with a newtype so I could override the
comparison function to sort by ys, then xs.

Now, as I'm committing this, I'm realizing that wasn't necessary,
because I ended up using an approach where the order of the dot points
didn't matter. I just loop over every point within bounds and print a
different character depending on if it's a dot or not. Oh well.