2021-D14A More exponential growth!
So I did the silly thing and generated the string again. I even had
figured out that the number of pairs in the string after n steps is m *
n^2 where m is the number of pairs in the starting string. I just didn't
actually think about what that implied about part 2.