2021-D14B Doing what I should have done in part A
Ironically this approach is one of the first ones that I thought of when
considering how to solve the problem directly in the first part. Keep a
map from pair to number of times that pair appears in the string, then
"expand" the string by looking up the new character between each pair
and replacing that entry with two new entries for each new pair. Then at
the end, flatten the map of pairs -> counts to a map of chars -> counts.
This requires only taking the last letter of each pair and adding one
to the count of the first letter in the string (since in the actual
string, each pair overlaps on both sides except the first and last, so
taking only the second of each pair and adding one for the first letter
will ensure no letters get counted twice).

Unfortunately, even after I realized the solution I went with for part a
wouldn't work, I still avoided doing this at first. I actually
implemented a different way of calculating the map of counts
recursively, which didn't work because there is just too many operations
there. Lots of time wasted. Oh well, hopefully lesson learned this time.