2021-D15A Finding the cheapest path
So I'm not sure if there's an even better way to do this one, there
probably is, but I remembered from some algorithms class I once took
that the A* algorithm is both better time-wise than a brute-force
approach, and guaranteed to get the least-cost path as long as the
heuristic (the h(n) function) follows the rule that h(n) is admissible
(i.e. it will never overestimate the actual least-cost from n to the
goal). In this case, a fairly trivial (if perhaps non-optimal) h(n) is
simply the orthogonal distance from n to the goal, as that is the
minimum number of positions the path must pass through, and each
position is guaranteed to have a cost >= 1.

There is probably some implementation of the A* algorithm for Haskell,
but I figured I should probably implement it myself, though it mainly
consisted of translating the pseudocode from Wikipedia as directly as
possible.