2021-D15B Bigger caves, bigger problems
I think using A* was the right choice. The expanded problem does take a
noticeable few seconds to run, but I suspect a brute-force approach
would be totally unfeasible. Of course, there may be an even better
algorithm that I just am completely missing, but if it works, it works.

Also, implementing the cave expansion algorithm was pretty annoying.
Lots of off-by-one errors and other mistakes.