2021-D16A Twiddling BITS
This was fun. The existence of the State monad really saved me here,
because otherwise it would have been a real pain managing the remaining
bits left to parse. I mean, it probably would have looked a lot like it
does now except without the conveniences of the "do" syntax. Which
admittedly is a lot of convenience.

Also, the existence of two standards for defining how many sub-packets
you have is just silly. Pick a lane!

Also I'm starting to really feel the pain of doing everything in a
single script. Particularly not having any kind of framework for tests.
Up to now I've just been manually running the sample input since in most
cases there was just one, but in this one there were quite a few small
samples that would have benefited by being in a test suite.
Unfortunately, the stack scripting feature doesn't allow for multi-file
scripts, so to fix this I'd have to convert it to a more standard
Haskell project. Maybe I'll find some time to do that, maybe I won't.