2021-D17A Going ballistic
The hardest part of this was probably figuring out the conditions for
terminating the search (i.e. when it's impossible for the torpedo to hit
the target because it's past it). To be honest I didn't do it
completely, my code very much relies on the target being in the positive
xs and negative ys.

It's also pretty far from elegant, something about this problem feels
difficult to write elegant code off the cuff.