2021-D18A Sailfish addition
Whew. This was a slog. First I had to parse a recursive language, which
is fun (good thing I had practice with the state monad). Then I needed
to implement a zipper for the recursive data structure I parsed it into,
so I could manipulate it. THEN I needed to implement a bunch of extra
operations on the zipper, such as "find the sibling to the left/right of
this node, modify that sibling, and return a zipper at the original
position (i.e. the current node) but with the modified sibling". THEN I
needed to actually implement the operations in the question (explode,
split, reduce, add, magnitude). Plus of course I had a bunch of test
cases provided by the question but no testing framework.

Overall I think the code could be improved, particularly the logic
around the explode and split operations seems both inelegant and
potentially inefficient. But it works, and it took almost no debugging
to get it that way.