2021-D20A Enhance. Enhance. Enhance!
This was pretty fun. I had some trouble with the getLed function because
at first I didn't realise the actual input has alternating infinite
fields of dark or light, so I just defaulted to off, then when I tried
to fix that I used the Point comparison to check the bounds, which
wasn't correct.