2021-D20B ENHAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAANCE
Easy to implement. I added a utility function to just re-run a function
on its own output a number of times, it seems like it might be
convenient.

This code is kind of slow. It takes almost a minute to run on the real
input, which obviously isn't ideal, but it's also workable, so I won't
bother trying to improve it.