2021-D21B Worlds upon worlds
So unfortunately some of my assumptions from the first part didn't pan
out, and I wasn't able to re-use a bunch of it for this part. But in the
end I did at least manage to re-use the Player struct, and one or two
functions related to that, so that's good.

The key here was realising that each player has at most 10 * 21 possible
states (the current position and score). That means, with only two
players, the max number of active game states is a mere 44100. Totally
small enough to keep track of each state along with how many universes
are currently in that state. Then, as I find states where a player has
won, I prune those and add it to the winning counts, so in the end I
should have 0 universes with ongoing games and the final tally of wins
for each player. Toss in a little uncertainty with the list monad, and
voila, a quantum game computer. Though it takes a few seconds to run,
and would probably die pretty quickly if we increased the parameters at
all.