2021-D22B More 3D manipulation, yayyyyy...
I thought there might be two different approaches to this.

The first and sort of more conceptually straightforward approach is to
maintain a list of discrete regions that represent all the "on" cubes at
that time.  Applying a step would mean applying the step to all the
individual regions, and combine the results into a new list. Basically
if the step is "on", then add a bunch of rectangular prisms that
together add up to the area not already covered by the currently on
cubes. If the step is "off" then break up the existing regions into
smaller ones and remove the parts that are being turned off.

The second approach would be to try to just keep a count of the
currently on cubes, and on each step try to update it by figuring out
how much of the region for the current step actually applies (or
alternatively, try to add and subtract extra values to account for
intersections with previous steps).

I tried to figure out the second approach for a while because I found it
pretty easy to come up with a simple way to calculate the intersection
of two regions, and I was hoping I could compose that in various ways to
keep an accurate count, and the prospect of splitting up the regions for
the first approach was pretty daunting. However, I ultimately decided
that figuring out the correct bunch of intersecting regions and whether
to add or subtract their volume from the count would be extremely
difficult in terms of complexity, and also possibly extremely
inefficient, as I was finding that the number of intersections to
consider was looking to increase combinatorially with the number of
prior steps.

Of course, the reason I tried so hard to avoid the first approach (which
I ended up doing anyway) was because I didn't want to deal with figuring
out the best way to split up a region into sub-regions that excluded
some other sub-region. I eventually worked out a way that let me do that
without having to figure out where the excluded sub-region was inside
the greater region (i.e. which corner it was in, or if it was
free-floating), by doing the equivalent logic for each axis, then
combining all the side segments into different sub-regions, and
filtering out the excluded one. This results in strictly more
sub-regions than necessary, but it greatly simplifies the code, so I
think it's worth it.

The final solution is still a bit slow, but no more so than most of the
other recent problems, so good enough.